Add a utility function to exec in shimdiag for cri-containerd tests
