+++ cmd/shimdiag/list.go
@@ -6,6 +6,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/Microsoft/hcsshim/internal/appargs"
+	"github.com/Microsoft/hcsshim/internal/shimdiag"
 	"github.com/urfave/cli"
 )
 
@@ -22,7 +23,7 @@ var listCommand = cli.Command{
 	Before: appargs.Validate(),
 	Action: func(ctx *cli.Context) error {
 		pids := ctx.Bool("pids")
-		shims, err := findShims("")
+		shims, err := shimdiag.FindShims("")
 		if err != nil {
 			return err
 		}
